requirements.txt
torch>=1.7.0
diffusers>=0.11.1
transformers>=4.25.1
accelerate>=0.15.0
Pillow>=9.0.0
numpy>=1.20.0